+++ components/src/test/java/com/mercadolibre/android/andesui/thumbnail/AndesThumbnailConfigurationTest.kt
@@ -19,7 +19,7 @@ import org.robolectric.RuntimeEnvironment
 import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
-@Config(sdk = [Build.VERSION_CODES.LOLLIPOP])
+@Config(sdk = [Build.VERSION_CODES.KITKAT, Build.VERSION_CODES.LOLLIPOP])
 class AndesThumbnailConfigurationTest {
 
     private var context = RuntimeEnvironment.application
@@ -27,72 +27,307 @@ class AndesThumbnailConfigurationTest {
     private val configFactory = spy(AndesThumbnailConfigurationFactory)
     private lateinit var attrs: AndesThumbnailAttrs
 
+    // MARK - Enabled Tests
+
     @Test
     fun `Thumbnail, Loud, Neutral, Icon, Enabled, background color`() {
+        // When
         val drawable = ContextCompat.getDrawable(context, R.drawable.andes_pagar_y_cobrar_efectivo_24)
-        attrs = AndesThumbnailAttrs(R.color.andes_orange_800.toAndesColor(), AndesThumbnailHierarchy.LOUD,
-                drawable!!, AndesThumbnailType.ICON, AndesThumbnailSize.SIZE_48, AndesThumbnailState.ENABLED)
+        attrs = AndesThumbnailAttrs(
+                R.color.andes_orange_800.toAndesColor(),
+                AndesThumbnailHierarchy.LOUD,
+                drawable!!,
+                AndesThumbnailType.ICON,
+                AndesThumbnailSize.SIZE_48,
+                AndesThumbnailState.ENABLED)
+
+        // Then
         val config = configFactory.create(context, attrs)
+
+        // Verify
         assertEquals(R.color.andes_orange_800.toAndesColor(), config.backgroundColor)
-        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
+        assertEquals(R.color.andes_gray_070_solid.toAndesColor(), config.borderColor)
+        assertEquals(AndesThumbnailHierarchy.LOUD.hierarchy.hasBorder(), config.hasBorder)
+        assertEquals(R.color.andes_white.toAndesColor(), config.iconColor)
+        assertEquals(whichIconSize(AndesThumbnailType.ICON, AndesThumbnailSize.SIZE_48), config.iconSize)
         assertEquals(drawable, config.image)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.cornerRadius)
+        assertEquals(whichImageType(AndesThumbnailType.ICON), config.isImageType)
+        assertEquals(true, config.hasTint)
     }
 
     @Test
     fun `Thumbnail, Quiet, Neutral, Icon, Enabled, background color`() {
+        // When
         val drawable = ContextCompat.getDrawable(context, R.drawable.andes_pagar_y_cobrar_efectivo_24)
-        val expectedColor = R.color.andes_orange_500.toAndesColor()
-        expectedColor.alpha = 0.1f
-        attrs = AndesThumbnailAttrs(R.color.andes_orange_500.toAndesColor(), AndesThumbnailHierarchy.QUIET,
-            drawable!!, AndesThumbnailType.ICON,
-            AndesThumbnailSize.SIZE_48, AndesThumbnailState.ENABLED)
+        val expectedIconColor = R.color.andes_orange_500.toAndesColor()
+        val expectedBackgrounColor = R.color.andes_orange_500.toAndesColor().apply {
+            alpha = 0.1f
+        }
+
+        attrs = AndesThumbnailAttrs(
+                R.color.andes_orange_500.toAndesColor(),
+                AndesThumbnailHierarchy.QUIET,
+                drawable!!,
+                AndesThumbnailType.ICON,
+                AndesThumbnailSize.SIZE_48,
+                AndesThumbnailState.ENABLED)
+
+        // Then
         val config = configFactory.create(context, attrs)
-        assertEquals(expectedColor, config.backgroundColor)
-        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
+
+        // Verify
+        assertEquals(expectedBackgrounColor, config.backgroundColor)
+        assertEquals(R.color.andes_gray_070_solid.toAndesColor(), config.borderColor)
+        assertEquals(AndesThumbnailHierarchy.QUIET.hierarchy.hasBorder(), config.hasBorder)
+        assertEquals(expectedIconColor, config.iconColor)
+        assertEquals(whichIconSize(AndesThumbnailType.ICON, AndesThumbnailSize.SIZE_48), config.iconSize)
         assertEquals(drawable, config.image)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.cornerRadius)
+        assertEquals(whichImageType(AndesThumbnailType.ICON), config.isImageType)
+        assertEquals(true, config.hasTint)
     }
 
     @Test
     fun `Thumbnail, Default, Neutral, Icon, Enabled, background color`() {
+        // When
         val drawable = ContextCompat.getDrawable(context, R.drawable.andes_pagar_y_cobrar_efectivo_24)
-        attrs = AndesThumbnailAttrs(R.color.andes_orange_800.toAndesColor(), AndesThumbnailHierarchy.DEFAULT,
-            drawable!!, AndesThumbnailType.ICON, AndesThumbnailSize.SIZE_48, AndesThumbnailState.ENABLED)
+        attrs = AndesThumbnailAttrs(
+                R.color.andes_orange_800.toAndesColor(),
+                AndesThumbnailHierarchy.DEFAULT,
+                drawable!!,
+                AndesThumbnailType.ICON,
+                AndesThumbnailSize.SIZE_48,
+                AndesThumbnailState.ENABLED)
+
+        // Then
         val config = configFactory.create(context, attrs)
+
+        // Verify
         assertEquals(R.color.andes_white.toAndesColor(), config.backgroundColor)
+        assertEquals(R.color.andes_gray_070_solid.toAndesColor(), config.borderColor)
+        assertEquals(AndesThumbnailHierarchy.DEFAULT.hierarchy.hasBorder(), config.hasBorder)
+        assertEquals(R.color.andes_gray_800_solid.toAndesColor(), config.iconColor)
+        assertEquals(whichIconSize(AndesThumbnailType.ICON, AndesThumbnailSize.SIZE_48), config.iconSize)
+        assertEquals(drawable, config.image)
         assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
-        assertEquals(drawable!!, config.image)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.cornerRadius)
+        assertEquals(whichImageType(AndesThumbnailType.ICON), config.isImageType)
+        assertEquals(true, config.hasTint)
     }
 
     @Test
-    fun `Thumbnail, Loud, Neutral, Icon, Disabled, background color`() {
+    fun `Thumbnail, Default, Neutral, Image Circle, Enabled, background color`() {
+        // When
         val drawable = ContextCompat.getDrawable(context, R.drawable.andes_pagar_y_cobrar_efectivo_24)
-        attrs = AndesThumbnailAttrs(R.color.andes_orange_800.toAndesColor(), AndesThumbnailHierarchy.LOUD,
-            drawable!!, AndesThumbnailType.ICON, AndesThumbnailSize.SIZE_48, AndesThumbnailState.DISABLED)
+        attrs = AndesThumbnailAttrs(
+                R.color.andes_orange_800.toAndesColor(),
+                AndesThumbnailHierarchy.DEFAULT,
+                drawable!!,
+                AndesThumbnailType.IMAGE_CIRCLE,
+                AndesThumbnailSize.SIZE_48,
+                AndesThumbnailState.ENABLED)
+
+        // Then
         val config = configFactory.create(context, attrs)
-        assertEquals(R.color.andes_gray_100_solid.toAndesColor(), config.backgroundColor)
+
+        // Verify
+        assertEquals(R.color.andes_white.toAndesColor(), config.backgroundColor)
+        assertEquals(R.color.andes_gray_070_solid.toAndesColor(), config.borderColor)
+        assertEquals(AndesThumbnailHierarchy.DEFAULT.hierarchy.hasBorder(), config.hasBorder)
+        assertEquals(R.color.andes_gray_800_solid.toAndesColor(), config.iconColor)
+        assertEquals(whichIconSize(AndesThumbnailType.IMAGE_CIRCLE, AndesThumbnailSize.SIZE_48), config.iconSize)
+        assertEquals(drawable, config.image)
         assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
-        assertEquals(drawable!!, config.image)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.cornerRadius)
+        assertEquals(whichImageType(AndesThumbnailType.IMAGE_CIRCLE), config.isImageType)
+        assertEquals(false, config.hasTint)
+    }
+
+    @Test
+    fun `Thumbnail, Loud, Neutral, Image Square, Enabled, background color`() {
+        // When
+        val drawable = ContextCompat.getDrawable(context, R.drawable.andes_pagar_y_cobrar_efectivo_24)
+        attrs = AndesThumbnailAttrs(
+                R.color.andes_orange_800.toAndesColor(),
+                AndesThumbnailHierarchy.LOUD,
+                drawable!!,
+                AndesThumbnailType.IMAGE_SQUARE,
+                AndesThumbnailSize.SIZE_48,
+                AndesThumbnailState.ENABLED)
+
+        // Then
+        val config = configFactory.create(context, attrs)
+
+        // Verify
+        assertEquals(R.color.andes_white.toAndesColor(), config.backgroundColor)
+        assertEquals(R.color.andes_gray_070_solid.toAndesColor(), config.borderColor)
+        assertEquals(AndesThumbnailHierarchy.DEFAULT.hierarchy.hasBorder(), config.hasBorder)
+        assertEquals(R.color.andes_gray_800_solid.toAndesColor(), config.iconColor)
+        assertEquals(whichIconSize(AndesThumbnailType.IMAGE_SQUARE, AndesThumbnailSize.SIZE_48), config.iconSize)
+        assertEquals(drawable, config.image)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.radiusSize(context), config.cornerRadius)
+        assertEquals(whichImageType(AndesThumbnailType.IMAGE_SQUARE), config.isImageType)
+        assertEquals(false, config.hasTint)
+    }
+
+    // MARK - Disabled Tests
+
+    @Test
+    fun `Thumbnail, Loud, Neutral, Icon, Disabled, background color`() {
+        // When
+        val drawable = ContextCompat.getDrawable(context, R.drawable.andes_pagar_y_cobrar_efectivo_24)
+        attrs = AndesThumbnailAttrs(
+                R.color.andes_orange_800.toAndesColor(),
+                AndesThumbnailHierarchy.LOUD,
+                drawable!!,
+                AndesThumbnailType.ICON,
+                AndesThumbnailSize.SIZE_48,
+                AndesThumbnailState.DISABLED)
+
+        // Then
+        val config = configFactory.create(context, attrs)
+
+        // Verify
+        assertEquals(R.color.andes_gray_100_solid.toAndesColor(), config.backgroundColor)
+        assertEquals(R.color.andes_gray_070_solid.toAndesColor(), config.borderColor)
+        assertEquals(AndesThumbnailHierarchy.LOUD.hierarchy.hasBorder(), config.hasBorder)
+        assertEquals(R.color.andes_gray_250_solid.toAndesColor(), config.iconColor)
+        assertEquals(whichIconSize(AndesThumbnailType.ICON, AndesThumbnailSize.SIZE_48), config.iconSize)
+        assertEquals(drawable, config.image)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.cornerRadius)
+        assertEquals(whichImageType(AndesThumbnailType.ICON), config.isImageType)
+        assertEquals(true, config.hasTint)
     }
 
     @Test
     fun `Thumbnail, Quiet, Neutral, Icon, Disabled, background color`() {
+        // When
         val drawable = ContextCompat.getDrawable(context, R.drawable.andes_pagar_y_cobrar_efectivo_24)
-        attrs = AndesThumbnailAttrs(R.color.andes_orange_500.toAndesColor(), AndesThumbnailHierarchy.QUIET,
-                drawable!!, AndesThumbnailType.ICON, AndesThumbnailSize.SIZE_48, AndesThumbnailState.DISABLED)
+        attrs = AndesThumbnailAttrs(
+                R.color.andes_orange_500.toAndesColor(),
+                AndesThumbnailHierarchy.QUIET,
+                drawable!!,
+                AndesThumbnailType.ICON,
+                AndesThumbnailSize.SIZE_48,
+                AndesThumbnailState.DISABLED)
+
+        // Then
         val config = configFactory.create(context, attrs)
+
+        // Verify
         assertEquals(R.color.andes_gray_100_solid.toAndesColor(), config.backgroundColor)
-        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
+        assertEquals(R.color.andes_gray_070_solid.toAndesColor(), config.borderColor)
+        assertEquals(AndesThumbnailHierarchy.QUIET.hierarchy.hasBorder(), config.hasBorder)
+        assertEquals(R.color.andes_gray_250_solid.toAndesColor(), config.iconColor)
+        assertEquals(whichIconSize(AndesThumbnailType.ICON, AndesThumbnailSize.SIZE_48), config.iconSize)
         assertEquals(drawable, config.image)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.cornerRadius)
+        assertEquals(whichImageType(AndesThumbnailType.ICON), config.isImageType)
+        assertEquals(true, config.hasTint)
     }
 
     @Test
     fun `Thumbnail, Default, Neutral, Icon, Disabled, background color`() {
+        // When
         val drawable = ContextCompat.getDrawable(context, R.drawable.andes_pagar_y_cobrar_efectivo_24)
-        attrs = AndesThumbnailAttrs(R.color.andes_orange_800.toAndesColor(), AndesThumbnailHierarchy.DEFAULT,
-            drawable!!, AndesThumbnailType.ICON, AndesThumbnailSize.SIZE_48, AndesThumbnailState.DISABLED)
+        attrs = AndesThumbnailAttrs(
+                R.color.andes_orange_800.toAndesColor(),
+                AndesThumbnailHierarchy.DEFAULT,
+                drawable!!,
+                AndesThumbnailType.ICON,
+                AndesThumbnailSize.SIZE_48,
+                AndesThumbnailState.DISABLED)
+
+        // Then
         val config = configFactory.create(context, attrs)
+
+        // Verify
         assertEquals(R.color.andes_white.toAndesColor(), config.backgroundColor)
-        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
+        assertEquals(R.color.andes_gray_070_solid.toAndesColor(), config.borderColor)
+        assertEquals(AndesThumbnailHierarchy.DEFAULT.hierarchy.hasBorder(), config.hasBorder)
+        assertEquals(R.color.andes_gray_250_solid.toAndesColor(), config.iconColor)
+        assertEquals(whichIconSize(AndesThumbnailType.ICON, AndesThumbnailSize.SIZE_48), config.iconSize)
         assertEquals(drawable, config.image)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.cornerRadius)
+        assertEquals(whichImageType(AndesThumbnailType.ICON), config.isImageType)
+        assertEquals(true, config.hasTint)
     }
+
+    @Test
+    fun `Thumbnail, Default, Neutral, Image Circle, Disabled, background color`() {
+        // When
+        val drawable = ContextCompat.getDrawable(context, R.drawable.andes_pagar_y_cobrar_efectivo_24)
+        attrs = AndesThumbnailAttrs(
+                R.color.andes_orange_800.toAndesColor(),
+                AndesThumbnailHierarchy.DEFAULT,
+                drawable!!,
+                AndesThumbnailType.IMAGE_CIRCLE,
+                AndesThumbnailSize.SIZE_48,
+                AndesThumbnailState.DISABLED)
+
+        // Then
+        val config = configFactory.create(context, attrs)
+
+        // Verify
+        assertEquals(R.color.andes_white.toAndesColor(), config.backgroundColor)
+        assertEquals(R.color.andes_gray_070_solid.toAndesColor(), config.borderColor)
+        assertEquals(AndesThumbnailHierarchy.DEFAULT.hierarchy.hasBorder(), config.hasBorder)
+        assertEquals(R.color.andes_gray_250_solid.toAndesColor(), config.iconColor)
+        assertEquals(whichIconSize(AndesThumbnailType.IMAGE_CIRCLE, AndesThumbnailSize.SIZE_48), config.iconSize)
+        assertEquals(drawable, config.image)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.cornerRadius)
+        assertEquals(whichImageType(AndesThumbnailType.IMAGE_CIRCLE), config.isImageType)
+        assertEquals(false, config.hasTint)
+    }
+
+    @Test
+    fun `Thumbnail, Loud, Neutral, Image Square, Disabled, background color`() {
+        // When
+        val drawable = ContextCompat.getDrawable(context, R.drawable.andes_pagar_y_cobrar_efectivo_24)
+        attrs = AndesThumbnailAttrs(
+                R.color.andes_orange_800.toAndesColor(),
+                AndesThumbnailHierarchy.LOUD,
+                drawable!!,
+                AndesThumbnailType.IMAGE_SQUARE,
+                AndesThumbnailSize.SIZE_48,
+                AndesThumbnailState.DISABLED)
+
+        // Then
+        val config = configFactory.create(context, attrs)
+
+        // Verify
+        assertEquals(R.color.andes_white.toAndesColor(), config.backgroundColor)
+        assertEquals(R.color.andes_gray_070_solid.toAndesColor(), config.borderColor)
+        assertEquals(AndesThumbnailHierarchy.DEFAULT.hierarchy.hasBorder(), config.hasBorder)
+        assertEquals(R.color.andes_gray_250_solid.toAndesColor(), config.iconColor)
+        assertEquals(whichIconSize(AndesThumbnailType.IMAGE_SQUARE, AndesThumbnailSize.SIZE_48), config.iconSize)
+        assertEquals(drawable, config.image)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.diameter(context), config.size)
+        assertEquals(AndesThumbnailSize.SIZE_48.size.radiusSize(context), config.cornerRadius)
+        assertEquals(whichImageType(AndesThumbnailType.IMAGE_SQUARE), config.isImageType)
+        assertEquals(false, config.hasTint)
+    }
+
+    // MARK - API Level Helpers
+
+    private fun whichIconSize(
+        icon: AndesThumbnailType,
+        size: AndesThumbnailSize
+    ): Int = if (icon != AndesThumbnailType.ICON && Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                size.size.diameter(context).toInt()
+            } else {
+                size.size.iconSize(context).toInt()
+            }
+
+    private fun whichImageType(
+        icon: AndesThumbnailType
+    ): Boolean = icon != AndesThumbnailType.ICON && Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP
 }