+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 # v3.10.1
 ## 🛠 Fixes
 - Andes Tooltip with showcase and pixel perfect issues resolved. | Author: [@RaelyxCordero](https://github.com/RaelyxCordero)
+- Andes Message Thumbnail without bottom margin
 - Support statusbar with variable height in CoachmarkScrolless
 - Fix color of AndesCheckBox to andes_accent_color_500